+++ src/components/RecommendedGallery.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import { MemoryRouter } from "react-router-dom";
-import { LikeContextProvider } from "../context/LikeContext";
+import { LikeContextProvider } from "../hooks/useLikeContext";
 import RecommendedGallery from "./RecommendedGallery";
 
 describe("RecommendedGallery component", () => {
